Converts lfm rating enum to IntDef

+++ app/src/main/kotlin/com/kelsos/mbrc/networking/DiscoveryStop.kt
@@ -1,13 +1,13 @@
 package com.kelsos.mbrc.networking
 
-import android.support.annotation.IntDef
+import androidx.annotation.IntDef
 
 object DiscoveryStop {
   const val NO_WIFI = 1
   const val NOT_FOUND = 2
   const val COMPLETE = 3
 
-  @IntDef(NO_WIFI.toLong(), NOT_FOUND.toLong(), COMPLETE.toLong())
+  @IntDef(NO_WIFI, NOT_FOUND, COMPLETE)
   @Retention(AnnotationRetention.SOURCE)
   annotation class Reason
 }
